Created a thread for Accelerometer and made myThreads file
The myThreads file is used to hold and control all the working threads and also holds the shutdown system of the program

+++ hal/include/hal/accelerometer.h
@@ -15,9 +15,9 @@ class Accelerometer{
         int getZ();
         void updateAccValues();
     private:
-        int16_t x;
-        int16_t y;
-        int16_t z;
+        float x;
+        float y;
+        float z;
         int i2cFileDesc;
         unsigned char regAddr;
 };
